+++ django_enum/__init__.py
@@ -48,4 +48,4 @@ __title__ = 'Django Enum'
 __version__ = '.'.join(str(i) for i in VERSION)
 __author__ = 'Brian Kohan'
 __license__ = 'MIT'
-__copyright__ = 'Copyright 2022-2023 Brian Kohan'
+__copyright__ = 'Copyright 2022-2024 Brian Kohan'